# src/app/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MeasureX API",
    "description": "Construction takeoff and measurement platform API",
    "version": "1.0.0",
    "contact": {
      "name": "MeasureX Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API base"
    }
  ],
...
  "paths": {
    "/api/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List all projects",
        "operationId": "listProjects",
        "responses": {
          "200": {
            "description": "List of projects with thumbnails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/Project"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "thumbnail": {
                                "type": "string",
                                "nullable": true
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create a new project",
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Building A — Floor Plan"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/Project"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get project with full hydrated state",
        "operationId": "getProject",
        "description": "Returns the project with all classifications, polygons, scale, pages,\nand thumbnail bundled in a single round-trip for client hydration.\n",
        "responses": {
          "200": {
            "description": "Full project state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Project"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "thumbnail": {
                              "type": "string",
                              "nullable": true
                            },
                            "state": {
                              "type": "object",
                              "properties": {
                                "classifications": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/Classification"
                                  }
                                },
                                "polygons": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/Polygon"
                                  }
                                },
                                "scale": {
                                  "$ref": "#/components/schemas/ScaleConfig"
                                },
                                "scales": {
                                  "type": "object",
                                  "additionalProperties": {
                                    "$ref": "#/components/schemas/ScaleConfig"
                                  }
                                },
                                "currentPage": {
                                  "type": "integer",
                                  "example": 1
                                },
                                "totalPages": {
                                  "type": "integer"
                                },
                                "sheetNames": {
                                  "type": "object",
                                  "additionalProperties": {
                                    "type": "string"
                                  }
                                },
                                "drawingSets": {
                                  "type": "object",
                                  "additionalProperties": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid project ID"
          },
          "404": {
            "description": "Project not found"
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Update project (autosave)",
        "operationId": "updateProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "state": {
                    "type": "object",
                    "properties": {
                      "scale": {
                        "$ref": "#/components/schemas/ScaleConfig"
                      },
                      "totalPages": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/Project"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Partial update (thumbnail, name, metadata)",
        "operationId": "patchProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "thumbnail": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patched project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/Project"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete a project",
        "operationId": "deleteProject",
        "responses": {
          "200": {
            "description": "Deletion confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{id}/scale": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "get": {
        "tags": [
          "Scale"
        ],
        "summary": "Get project scale",
        "operationId": "getScale",
        "responses": {
          "200": {
            "description": "Current scale configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scale": {
                      "$ref": "#/components/schemas/ScaleConfig",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Scale"
        ],
        "summary": "Set project scale",
        "operationId": "setScale",
        "description": "Sets the drawing scale. This is critical for converting pixel measurements\nto real-world units. Broadcasts `scale:updated` via SSE.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "pixelsPerUnit",
                  "unit"
                ],
                "properties": {
                  "pixelsPerUnit": {
                    "type": "number",
                    "description": "How many pixels equal one real-world unit",
                    "example": 96
                  },
                  "unit": {
                    "type": "string",
                    "description": "Real-world unit (ft, in, m, etc.)",
                    "example": "ft"
                  },
                  "label": {
                    "type": "string",
                    "default": "Custom",
                    "example": "1/4\" = 1'-0\""
                  },
                  "source": {
                    "type": "string",
                    "default": "manual",
                    "enum": [
                      "manual",
                      "auto",
                      "ai"
                    ]
                  },
                  "pageNumber": {
                    "type": "integer",
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scale saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scale": {
                      "$ref": "#/components/schemas/ScaleConfig"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ProjectId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-zA-Z0-9_-]+$"
        },
        "description": "Project UUID"
      }
    },
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "totalPages": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Classification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "area",
              "linear",
              "count"
            ]
          },
          "color": {
            "type": "string"
          },
          "visible": {
            "type": "boolean"
          },
          "formula": {
            "type": "string",
            "nullable": true
          },
          "formulaUnit": {
            "type": "string",
            "nullable": true
          },
          "formulaSavedToLibrary": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Polygon": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "classificationId": {
            "type": "string",
            "format": "uuid"
          },
          "points": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "x": {
                  "type": "number"
                },
                "y": {
                  "type": "number"
                }
              }
            }
          },
          "pageNumber": {
            "type": "integer"
          },
          "area": {
            "type": "number"
          },
          "linearFeet": {
            "type": "number"
          },
          "isComplete": {
            "type": "boolean"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ScaleConfig": {
        "type": "object",
        "properties": {
          "pixelsPerUnit": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "pageNumber": {
            "type": "integer"
          },
          "confidence": {
            "type": "number"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}