# src/app/api/projects/[id]/route.ts
import { NextResponse } from 'next/server';
import { getProject, updateProject, deleteProject, initDataDir, getClassifications, getPolygons, getScale, setScale, getPages, getThumbnail } from '@/server/project-store';
import type { Classification, Polygon } from '@/lib/types';
import type { PageInfo } from '@/server/project-store';
import { ProjectIdSchema, ProjectUpdateSchema, validationError } from '@/lib/api-schemas';

export async function GET(_req: Request, { params }: { params: Promise<{ id: string }> }) {
  try {
    await initDataDir();
    const paramsResult = ProjectIdSchema.safeParse(await params);
    if (!paramsResult.success) return validationError(paramsResult.error);
    const { id } = paramsResult.data;
    const project = await getProject(id);
    if (!project) return NextResponse.json({ error: 'Project not found' }, { status: 404 });

    // Bundle full state so the client can hydrate in a single round-trip
    const [classifications, polygons, scale, pages, thumbnail] = await Promise.all([
      getClassifications(id).catch(() => [] as Classification[]),
      getPolygons(id).catch(() => [] as Polygon[]),
      getScale(id).catch(() => null),
      getPages(id).catch((e) => { console.error('getPages error:', e); return [] as PageInfo[]; }),
      getThumbnail(id).catch(() => null),
    ]);

    // totalPages: prefer stored value (project.totalPages), fall back to mx_pages count
    const totalPages = (project.totalPages && project.totalPages > 1)
      ? project.totalPages
      : (pages.length > 1 ? pages.length : (project.totalPages ?? 1));

    // Build sheetNames and drawingSets maps from stored page data
    const sheetNames: Record<number, string> = {};
    const drawingSets: Record<number, string> = {};
    for (const p of pages) {
      if (p.name) sheetNames[p.pageNum] = p.name;
      if (p.drawingSet) drawingSets[p.pageNum] = p.drawingSet;
    }

    return NextResponse.json({
      project: {
        ...project,
        thumbnail,
        state: {
          classifications,
          polygons,
          scale,
          scales: {},
          currentPage: 1,
          totalPages,
          sheetNames,
          drawingSets,
        },
      },
    });
  } catch (err: unknown) {
    console.error("[project route]", err);
    return NextResponse.json({ error: (err instanceof Error ? err.message : String(err)) }, { status: 500 });
  }
}

export async function PUT(req: Request, { params }: { params: Promise<{ id: string }> }) {
  try {
    await initDataDir();
    const paramsResult = ProjectIdSchema.safeParse(await params);
    if (!paramsResult.success) return validationError(paramsResult.error);
    const { id } = paramsResult.data;
    const body = await req.json();

    // If the body contains a `state` payload (autosave from client), persist
    // the scale from that state. Polygons/classifications are synced individually
    // via their own endpoints as they're created, so we only need to handle scale here.
    const state = body.state;
    if (state?.scale) {
      const s = state.scale;
      await setScale(id, {
        pixelsPerUnit: s.pixelsPerUnit,
        unit: s.unit,
        label: s.label || 'Custom',
        source: s.source || 'manual',
        pageNumber: s.pageNumber || 1,
        confidence: s.confidence,
      }).catch(() => null); // non-fatal
    }

    // Update project metadata (name, totalPages) if provided — extract only safe fields
    const metaPatch: { name?: string; totalPages?: number } = {};
    if (typeof body.name === 'string') metaPatch.name = body.name;
    if (typeof state?.totalPages === 'number' && state.totalPages > 0) {
      metaPatch.totalPages = state.totalPages;
    }
    const updated = await updateProject(id, metaPatch);
    return NextResponse.json({ project: updated });
  } catch (err: unknown) {
    console.error("[project route]", err);
    return NextResponse.json({ error: (err instanceof Error ? err.message : String(err)) }, { status: 500 });
  }
}

export async function PATCH(req: Request, { params }: { params: Promise<{ id: string }> }) {
  try {
    await initDataDir();
    const paramsResult = ProjectIdSchema.safeParse(await params);
    if (!paramsResult.success) return validationError(paramsResult.error);
    const { id } = paramsResult.data;
    const body = await req.json();
    const patch: Record<string, unknown> = {};
    if (typeof body.thumbnail === 'string') patch.thumbnail = body.thumbnail;
    if (typeof body.name === 'string') patch.name = body.name;
    if (body.metadata && typeof body.metadata === 'object') patch.metadata = body.metadata;
    const updated = await updateProject(id, patch as any);
    return NextResponse.json({ project: updated });
  } catch (err: unknown) {
    return NextResponse.json({ error: (err instanceof Error ? err.message : String(err)) }, { status: 500 });
  }
}

export async function DELETE(_req: Request, { params }: { params: Promise<{ id: string }> }) {
  try {
    await initDataDir();
    const paramsResult = ProjectIdSchema.safeParse(await params);
    if (!paramsResult.success) return validationError(paramsResult.error);
    const { id } = paramsResult.data;
    const ok = await deleteProject(id);
    return NextResponse.json({ ok });
  } catch (err: unknown) {
    console.error("[project route]", err);
    return NextResponse.json({ error: (err instanceof Error ? err.message : String(err)) }, { status: 500 });
  }
}